Clases principales terminadas. Clase UsuarioDAO terminada.

+++ src/main/java/com/example/gestordepedidoshibernate/domain/dao/DAO.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 
 public interface DAO<T> {
     public ArrayList<T> getAll();
-    public T get(Long id);
+    public T get(Integer id);
     public T save(T data);
     public void update(T data);
     public void delete(T data);
